NR-234826 - Missing DT Headers (#211)
* NR-234826 - Added PayloadHolder

Added a PayloadHolder class to enable the proper payload type to have its headers added prior to the request, and then by being added to the request object afterward.

* Ignore newState when checking the state of a task

newState is often passed in as nil, even though the task itself has been started and running. This is verified by checking the task's state itself.