NR-239832 - Reverted change that resulted in nil
The NSHTTPURLResponse that is assigned in this autorelease pool is referenced later in the method. If that is not assigned here, then it will be nil, which is perfectly fine to be, however, that will cause issues with parts of the agent which rely on the status code here for enabling or disabling the agent.