More Granular Reporting for Watch OS orientation
Instead of being hardcoded to "Landscape" like tvOS, watchOS now reports what wrist the device is on, whether the crown is on the left or right side of the device, and if the display has been autorotated. Autorotation would be the case if the user has flipped their wrist to show someone else the screen, for instance.